chore: 代码重构,将Mymap中的两个控制组件类抽出来放在单独的文件中
