+++ 记录.md
@@ -40,7 +40,7 @@
 - [x] 作者生平轨迹外国点处的markpoint绘制有问题,但不知道为什么又没了
 - [ ] **路书的具体功能实现: 途经点显示文字 路径 速度控制**
 - [ ] **右侧抽屉按钮功能**
-- [ ] **地图换颜色**
+- [x] **地图换颜色**
 - [ ] 课文详细信息再做一个页面
 - [ ] 小组思考,题目,注释翻译集成到文本阅读里 ?
 - [ ] 详情页也可以放这些
@@ -80,7 +80,6 @@
 # 本周计划:
 - [ ] **路书的具体功能实现: 途经点显示文字 路径 速度控制**
   > 路途中文字考虑加上起始点和原因
-  > 
   > 速度考虑等时
 
 - [ ] **右侧抽屉按钮功能**
@@ -91,10 +90,12 @@
 
 - [ ] 别的可视化图表
 
+  - **关于地图上的可视化工作需要多参考 百度地图 可视化图层api 这个其实不错的**
+
   - echars for react
   - 多做一些和时空信息相关的
 
-- [ ] **地图换颜色**
+- [x] **地图换颜色**
